Optimization changes
The individual cycle counters for movement are now by default 0 and return to 0 once a move was executed. Also, the cycle time limit gets incremented by 1 second everytime a move is completed (this is used to test that the functions work with different cycle limits during run time). 

In the old version the individual movement cycle counters would count up indefinitely.